python-backend/requirements.txt
# QuotePro Python Backend Dependencies

# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Environment variables
python-dotenv==1.0.1

# AI/ML
google-generativeai>=0.8.5
google-adk==1.20.0

# Database
supabase==2.10.0

# Data validation
pydantic==2.10.3
pydantic-settings==2.6.1

# File processing (CSV/Excel)
pandas==2.2.3
openpyxl==3.1.5

# HTTP requests
httpx>=0.26,<0.28
requests==2.32.3

# System monitoring
psutil==6.1.0

# Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
slowapi==0.1.9
bleach==6.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
